Modify the Attendance, Lines Written and Commits data
The data that is present in the students array has to be compatible
in terms of the key value pairs to be used to render charts. Therefore,
functions such as linesWrittenTemplate(), numberOfCommitsTemplate() have
been written to provide a standard template of data for the linesWritten
and numberOfCommits properties respectively. Functions such as
populateLinesWritten() and populateCommits() have been written in order
to automate the data generation. The function totalClassesHeld() return
the data regarding the total classes that have been held during the
months of January, February and March