Include storing user data in localStorage
In the previous version, the data was not persistent, which means that
as soon as the page was refreshed, the user data was lost. Hence, the
reducers have been modified in a way that when the login or logout action
is dispatched, the user data is stored in localStorage. The getUser method
is also modified in order to retrieve data from the localStorage instead
of the store.